Add a stub of the field element API.
This defines (but does not implement):

* conversion to and from the curve25519-dalek packed `Scalar` type;
* addition;
* subtraction;
* multiplication;
* zero & one constructors and Default;
* inversion;

The implementation in curve25519-dalek allows doing operations either in
Montgomery form or not in Montgomery form; I think that for this use-case, it
would be better to require that a `FieldElement` is *always* in Montgomery
form, so that the multiplication is always Montgomery multiplication.  We can
also probably save some time by allowing the representatives to lie in the
range [0, 2*l) instead of [0, l) as in the original implementation (cf. "Montgomery
Multiplication Needs No Final Subtractions").